Refactor core logic and improve code quality
- Extracted data sorting logic from `App.jsx` into pure functions in `src/utils/dataHandlers.js`.
- Moved `File` class and `fetchFile` utility to `src/utils/file.js` for better modularity.
- Refactored grade average calculation in `src/utils/gradesTools.js` to reduce code duplication.
- Modernized codebase by replacing `var` with `const`/`let`.
- Fixed various typos (e.g., `displfayMode` -> `displayMode`).
- Updated components to handle the refactored utilities.
- Verified build integrity with `npm run build`.
